requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
playsound>=1.3.0

# FastAPI 後端
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# LLM 支援（可選）
openai>=1.0.0  # 用於 OpenAI GPT 和 Whisper
anthropic>=0.7.0  # 用於 Claude
sentry-sdk>=1.39.0  # 錯誤追蹤

# 語音處理（可選）
# openai-whisper  # 本地 Whisper（如果需要）
# pyaudio  # 麥克風錄音（如果需要）

